@@ -1,4 +1,5 @@
-﻿using PSQuickAssets.Models;
+﻿using PSQuickAssets.Assets;
+using PSQuickAssets.Models;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
